test(AUser): fix test regression due to invalid partial query result assertion
this issue was caught in the demo environment daily test.
partial query can return multiple matches if the partial query term is not unique (not just the expected test user).
our test result should not be affected by already existing data in the test environment.
also we are not supposed to test backend behavior here, so we remove those tests